Add live activities example.
In this case there is an app for iOS 18 that shows how to implement
live activtities.

The app is divided in several pacakges and targets:

- LiveActivityDemo
  - sample iOS app
- DemoWidgets
  - widget bundle that defines the live activity
- LiveActivityContent
  - package where the Live Activity content is defined, to be reused
    easily between the iOS app, widget bundle and JSONPayload script
- LiveActivityUI
  - package that contains the UI of live activity visible in the widget
    bundle
- JSONPayload
  - package that defines a script to generate the push notification JSON
    content to start, update and end the live acitivty via push
    notifications.

The live activity can be started via push, using the [CloudKit Console](https://icloud.developer.apple.com/dashboard/notifications)
is the easiest way to send pushes to a device or a simulator.